requirements.txt
numpy
pandas
pvlib
matplotlib
numpy
scipy
pytz
Flask
Flask-SQLAlchemy
Pillow